# Conflicts:
#	project/src/main/kotlin/cga/exercise/game/Scene.kt

+++ project/src/main/kotlin/cga/exercise/components/geometry/Transformable.kt
@@ -1,6 +1,8 @@
 package cga.exercise.components.geometry
 
 import org.joml.Matrix4f
+import org.joml.Quaterniond
+import org.joml.Quaternionf
 import org.joml.Vector3f
 
 open class Transformable(private var modelMatrix: Matrix4f = Matrix4f(), var parent: Transformable? = null) {
@@ -28,8 +30,8 @@ open class Transformable(private var modelMatrix: Matrix4f = Matrix4f(), var par
      * @param yaw radiant angle around y-axis ccw
      * @param roll radiant angle around z-axis ccw
      */
-    fun rotate(pitch: Float, yaw: Double, roll: Float) {
-        modelMatrix.rotateXYZ(pitch, yaw.toFloat(), roll)
+    fun rotate(pitch: Float, yaw: Float, roll: Float) {
+        modelMatrix.rotateXYZ(pitch, yaw, roll)
     }
     /**
      * Rotates object around given rotation center.
